// vue-source/packages/reactivity/src/effect.d.ts
export type DebuggerEvent = {
    effect: Subscriber;
} & DebuggerEventExtraInfo;
export type DebuggerEventExtraInfo = {
    target: object;
    type: TrackOpTypes | TriggerOpTypes;
    key: any;
    newValue?: any;
    oldValue?: any;
    oldTarget?: Map<any, any> | Set<any>;
};
export interface DebuggerOptions {
    onTrack?: (event: DebuggerEvent) => void;
    onTrigger?: (event: DebuggerEvent) => void;
}
export interface ReactiveEffectOptions extends DebuggerOptions {
    scheduler?: EffectScheduler;
    allowRecurse?: boolean;
    onStop?: () => void;
}
export interface ReactiveEffectRunner<T = any> {
    (): T;
    effect: ReactiveEffect;
}
export declare let activeSub: Subscriber | undefined;
export declare enum EffectFlags {
    /**
     * ReactiveEffect only
     */
    ACTIVE = 1,
    RUNNING = 2,
    TRACKING = 4,
    NOTIFIED = 8,
    DIRTY = 16,
    ALLOW_RECURSE = 32,
    PAUSED = 64,
    EVALUATED = 128
}
/**
 * Subscriber is a type that tracks (or subscribes to) a list of deps.
 */
export interface Subscriber extends DebuggerOptions {
    /**
     * Head of the doubly linked list representing the deps
     * @internal
     */
    deps?: Link;
    /**
     * Tail of the same list
     * @internal
     */
    depsTail?: Link;
    /**
     * @internal
     */
    flags: EffectFlags;
    /**
     * @internal
     */
    next?: Subscriber;
    /**
     * returning `true` indicates it's a computed that needs to call notify
     * on its dep too
     * @internal
     */
    notify(): true | void;
}